# Add AnyHolder to improve blockImpl performance