--- a/bcos-framework/bcos-framework/protocol/Block.h
+++ b/bcos-framework/bcos-framework/protocol/Block.h
@@ -97,24 +97,24 @@ public:
     virtual uint64_t receiptsSize() const = 0;
 
     // for nonceList
-    virtual void setNonceList(RANGES::any_view<std::string> nonces) = 0;
-    virtual RANGES::any_view<std::string> nonceList() const = 0;
+    virtual void setNonceList(::ranges::any_view<std::string> nonces) = 0;
+    virtual ::ranges::any_view<std::string> nonceList() const = 0;
 
     virtual NonceListPtr nonces() const
     {
         return std::make_shared<NonceList>(
-            RANGES::iota_view<size_t, size_t>(0LU, transactionsSize()) |
-            RANGES::views::transform([this](uint64_t index) {
+            ::ranges::iota_view<size_t, size_t>(0LU, transactionsSize()) |
+            ::ranges::views::transform([this](uint64_t index) {
                 auto transaction = this->transaction(index);
                 return transaction->nonce();
             }) |
-            RANGES::to<NonceList>());
+            ::ranges::to<NonceList>());
     }
 
     virtual ViewResult<crypto::HashType> transactionHashes() const = 0;
-    virtual ViewResult<std::unique_ptr<TransactionMetaData>> transactionMetaDatas() const = 0;
-    virtual ViewResult<std::unique_ptr<Transaction>> transactions() const = 0;
-    virtual ViewResult<std::unique_ptr<TransactionReceipt>> receipts() const = 0;
+    virtual ViewResult<AnyTransactionMetaData> transactionMetaDatas() const = 0;
+    virtual ViewResult<AnyTransaction> transactions() const = 0;
+    virtual ViewResult<AnyTransactionReceipt> receipts() const = 0;
     bool operator<(const Block& block) const
     {
         return blockHeaderConst()->number() < block.blockHeaderConst()->number();