--- a/bcos-tars-protocol/bcos-tars-protocol/protocol/TransactionMetaDataImpl.h
+++ b/bcos-tars-protocol/bcos-tars-protocol/protocol/TransactionMetaDataImpl.h
@@ -25,7 +25,9 @@
 
 namespace bcostars::protocol
 {
-class TransactionMetaDataImpl : public bcos::protocol::TransactionMetaData
+class TransactionMetaDataImpl
+  : public bcos::protocol::TransactionMetaData,
+    public virtual bcos::MoveImpl<TransactionMetaDataImpl, bcos::protocol::TransactionMetaData>
 {
 public:
     using Ptr = std::shared_ptr<TransactionMetaDataImpl>;
@@ -35,6 +37,10 @@ public:
     TransactionMetaDataImpl(bcos::crypto::HashType hash, std::string to);
     explicit TransactionMetaDataImpl(std::function<bcostars::TransactionMetaData*()> inner);
     ~TransactionMetaDataImpl() override = default;
+    TransactionMetaDataImpl& operator=(const TransactionMetaDataImpl& _txMetaData) = delete;
+    TransactionMetaDataImpl& operator=(TransactionMetaDataImpl&& _txMetaData) = default;
+    TransactionMetaDataImpl(const TransactionMetaDataImpl& _txMetaData) = delete;
+    TransactionMetaDataImpl(TransactionMetaDataImpl&& _txMetaData) = default;
 
     bcos::crypto::HashType hash() const override;
     void setHash(bcos::crypto::HashType _hash) override;