----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
testByIndex/threads:1       39.0 ns         40.7 ns     17354891
testByIndex/threads:4        280 ns          292 ns      2363292
testByIndex/threads:8        458 ns          477 ns      1426560
testByRange/threads:1       33.0 ns         34.3 ns     20720358
testByRange/threads:4       36.8 ns         38.1 ns     17579140
testByRange/threads:8       42.9 ns         44.7 ns     15907032